refactor(user-service): split retry types
